FIX: Specify files relative to git root for README
The location of the  markdown files whos content is used to generate
the README are now specified relative to the git repository root,
rather than relative to the scripts directory. This means the
generate_readme.sh script can now be run from anywhere, not just
from inside the scripts directory.

Resolves #1