ENH: Add pre-commit hook that auto-generates README
As the README is generated from a number of files by the script
scripts/generate_readme.sh, it is easy to make changes to the component
files and subsequently forget to run the script. This hook, which
should be copied to .git/hooks/, runs the script before every commit,
ensuring the README is updated and staged without the user having to do
anything.

Resolves #2